refactor(styles): add dash as a configuration option
Previously a dash instance was created inside of the `createStyles()` function using the options
provided. Now, those configuration options are set on the dash instance itself. That instance is
provided to `createStyles()` through the `dash` property in options now.

BREAKING CHANGE: Dash options are no longer assignable to `createStyles()`. Now, you have to ceate a
dash instance with `createDash()` and provide it to the `dash` option in `createStyles()` options.