refactor(styles): rename all instances of "variables" to "tokens"
"tokens" is a more descriptive term because we're describing "design tokens" more so than we are
just "CSS variables". This puts us in the language of a design system vs just a CSS-in-JS library.

BREAKING CHANGE: All instances of "variables" in the API was renamed to "tokens", all instances of
"Variables" was renamed to "Tokens"